+++ src/stubllm/pytest_plugin/__init__.py
@@ -1,4 +1,4 @@
-"""pytest plugin for mockllm."""
+"""pytest plugin for stubllm."""
 
 from stubllm.pytest_plugin.plugin import MockLLMServerFixture, use_fixtures